renamed menu.js to meal.js
